requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
numpy==1.26.0
pytest==8.0.0